Move purgeFile to DatabaseUtils.ts
